🐛 :: Type missmatch오류 수정

+++ domain/src/test/java/com/semicolon/domain/usecase/school/SearchSchoolUseCaseUnitTest.kt
@@ -25,7 +25,7 @@ class SearchSchoolUseCaseUnitTest {
             SearchSchoolEntity(
                 schoolList = listOf(
                     SearchSchoolEntity.SchoolInfo(
-                        schoolId,
+                        schoolId.toLong(),
                         image,
                         schoolName
                     )
